+++ frontend/views/site/index.php
@@ -1,148 +1,24 @@
 <?php
 
 /* @var $this yii\web\View */
+use common\widgets\CardList\CardList;
+use domain\models\base\search\AuthItemSearch;
+use yii\helpers\Url;
+
+/* @var $modelSearch AuthItemSearch */
 
 $this->title = 'My Yii Application';
 ?>
 <div class="site-index">
-    <div class="row">
-        <div class="col-md-4">
-            <!-- Media, Title, and Description area -->
-            <div class="pmd-card pmd-card-default pmd-z-depth">
 
-                <!-- Card media -->
-                <div class="pmd-card-media">
-                    <img class="img-responsive"
-                         src="http://propeller.in/assets/images/profile-pic.png">
-                </div>
+    <?= CardList::widget([
+        'url' => Url::to(['site/test']),
+        'search' => [
+            'modelSearch' => $modelSearch,
+            'searchAttributeName' => 'description',
+        ],
+        'popularity' => true,
+    ])
+    ?>
 
-                <!-- Card header -->
-                <div class="pmd-card-title">
-                    <h2 class="pmd-card-title-text">Департамент здравоохранения Ханты-Мансийского автономного округа -
-                        Югры</h2>
-                    <span class="pmd-card-subtitle-text">Официальный сайт Департамента здравоохранения Ханты-Мансийского автономного округа - Югры</span>
-                </div>
-
-                <!-- Card action -->
-                <div class="pmd-card-actions">
-                    <a class="btn pmd-btn-flat pmd-ripple-effect btn-primary" href="<?= Yii::$app->urlManagerAdmin->createUrl(['/']) ?>">Перейти</a>
-                </div>
-            </div>
-        </div>
-        <div class="col-md-4">
-            <!-- Media, Title, and Description area -->
-            <div class="pmd-card pmd-card-default pmd-z-depth">
-
-                <!-- Card media -->
-                <div class="pmd-card-media">
-                    <img class="img-responsive"
-                         src="http://propeller.in/assets/images/profile-pic.png">
-                </div>
-
-                <!-- Card header -->
-                <div class="pmd-card-title">
-                    <h2 class="pmd-card-title-text">Департамент здравоохранения Ханты-Мансийского автономного округа -
-                        Югры</h2>
-                    <span class="pmd-card-subtitle-text">Официальный сайт Департамента здравоохранения Ханты-Мансийского автономного округа - Югры</span>
-                </div>
-
-                <!-- Card action -->
-                <div class="pmd-card-actions">
-                    <button class="btn pmd-btn-flat pmd-ripple-effect btn-primary" type="button">Перейти</button>
-                </div>
-            </div>
-        </div>
-        <div class="col-md-4">
-            <!-- Media, Title, and Description area -->
-            <div class="pmd-card pmd-card-default pmd-z-depth">
-
-                <!-- Card media -->
-                <div class="pmd-card-media">
-                    <img class="img-responsive"
-                         src="http://propeller.in/assets/images/profile-pic.png">
-                </div>
-
-                <!-- Card header -->
-                <div class="pmd-card-title">
-                    <h2 class="pmd-card-title-text">Департамент здравоохранения Ханты-Мансийского автономного округа -
-                        Югры</h2>
-                    <span class="pmd-card-subtitle-text">Официальный сайт Департамента здравоохранения Ханты-Мансийского автономного округа - Югры</span>
-                </div>
-
-                <!-- Card action -->
-                <div class="pmd-card-actions">
-                    <button class="btn pmd-btn-flat pmd-ripple-effect btn-primary" type="button">Перейти</button>
-                </div>
-            </div>
-        </div>
-        <div class="col-md-4">
-            <!-- Media, Title, and Description area -->
-            <div class="pmd-card pmd-card-default pmd-z-depth">
-
-                <!-- Card media -->
-                <div class="pmd-card-media">
-                    <img class="img-responsive"
-                         src="http://propeller.in/assets/images/profile-pic.png">
-                </div>
-
-                <!-- Card header -->
-                <div class="pmd-card-title">
-                    <h2 class="pmd-card-title-text">Департамент здравоохранения Ханты-Мансийского автономного округа -
-                        Югры</h2>
-                    <span class="pmd-card-subtitle-text">Официальный сайт Департамента здравоохранения Ханты-Мансийского автономного округа - Югры</span>
-                </div>
-
-                <!-- Card action -->
-                <div class="pmd-card-actions">
-                    <button class="btn pmd-btn-flat pmd-ripple-effect btn-primary" type="button">Перейти</button>
-                </div>
-            </div>
-        </div>
-        <div class="col-md-4">
-            <!-- Media, Title, and Description area -->
-            <div class="pmd-card pmd-card-default pmd-z-depth">
-
-                <!-- Card media -->
-                <div class="pmd-card-media">
-                    <img class="img-responsive"
-                         src="http://propeller.in/assets/images/profile-pic.png">
-                </div>
-
-                <!-- Card header -->
-                <div class="pmd-card-title">
-                    <h2 class="pmd-card-title-text">Департамент здравоохранения Ханты-Мансийского автономного округа -
-                        Югры</h2>
-                    <span class="pmd-card-subtitle-text">Официальный сайт Департамента здравоохранения Ханты-Мансийского автономного округа - Югры</span>
-                </div>
-
-                <!-- Card action -->
-                <div class="pmd-card-actions">
-                    <button class="btn pmd-btn-flat pmd-ripple-effect btn-primary" type="button">Перейти</button>
-                </div>
-            </div>
-        </div>
-        <div class="col-md-4">
-            <!-- Media, Title, and Description area -->
-            <div class="pmd-card pmd-card-default pmd-z-depth">
-
-                <!-- Card media -->
-                <div class="pmd-card-media">
-                    <img class="img-responsive"
-                         src="http://propeller.in/assets/images/profile-pic.png">
-                </div>
-
-                <!-- Card header -->
-                <div class="pmd-card-title">
-                    <h2 class="pmd-card-title-text">Департамент здравоохранения Ханты-Мансийского автономного округа -
-                        Югры</h2>
-                    <span class="pmd-card-subtitle-text">Официальный сайт Департамента здравоохранения Ханты-Мансийского автономного округа - Югры</span>
-                </div>
-
-                <!-- Card action -->
-                <div class="pmd-card-actions">
-                    <button class="btn pmd-btn-flat pmd-ripple-effect btn-primary" type="button">Перейти</button>
-                </div>
-            </div>
-        </div>
-    </div>
 </div>